requirements.dev.txt
asgiref==3.4.1
astor==0.8.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
cognitive-complexity==1.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.1
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
django-ipware==4.0.0
django-split-settings==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
eradicate==2.0.0
flake8==3.9.2
flake8-absolute-import==1.0
flake8-bugbear==21.4.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-django==1.1.2
flake8-eradicate==1.1.0
flake8-fixme==1.1.1
flake8-isort==4.0.0
flake8-mock==0.3
flake8-multiline-containers==0.0.18
flake8-mutable==1.2.0
flake8-pep3101==1.3.0
flake8-pie==0.6.1
flake8-plugin-utils==1.3.2
flake8-print==4.0.0
flake8-printf-formatting==1.1.2
flake8-pytest==1.3
flake8-pytest-style==1.5.0
flake8-quotes==3.3.0
flake8-simplify==0.14.1
flake8-todo==0.7
flake8-use-fstring==1.1
flake8-variables-names==0.0.4
idna==3.2
importlib-metadata==4.7.1
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.0.3
packaging==21.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.14
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
testfixtures==6.18.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
